Add minor changes in Repositories Add UserService with basic methods Add DictionaryService with basic methods Fix problem in Dictionary with @OneToMany words List

+++ src/main/java/org/example/dictionary/repository/DictionaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DictionaryRepository extends JpaRepository<Dictionary, Integer> {
+public interface DictionaryRepository extends JpaRepository<Dictionary, Long> {
 }
